refactor: move session delegates to dlg/session/ subfolder
