+++ src/dlg/session/SubmitAnswer.ts
@@ -1,8 +1,8 @@
 import { Request } from "express";
 import { TotoDelegate, UserContext, ValidationError } from "totoms";
-import { ControllerConfig } from "../Config";
-import { SessionAnswer } from "../model/Session";
-import { SessionsStore } from "../store/SessionsStore";
+import { ControllerConfig } from "../../Config";
+import { SessionAnswer } from "../../model/Session";
+import { SessionsStore } from "../../store/SessionsStore";
 
 export class SubmitAnswer extends TotoDelegate<SubmitAnswerRequest, SubmitAnswerResponse> {
